refactor: Move function to trajectory package
